// src/sdk/generated/models/TopicUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
/**
 * Schema for updating a topic
 */
export type TopicUpdate = {
  name?: string | null;
  code?: string | null;
  description?: string | null;
  content?: string | null;
  video_url?: string | null;
  document_url?: string | null;
  parent_id?: string | null;
  order?: number | null;
  estimated_time_minutes?: number | null;
  difficulty_level?: string | null;
  is_active?: boolean | null;
};